+++ SCME.Types/ProhibitionVoltage/TypeProhibitionVoltage.cs
@@ -27,24 +27,24 @@ namespace SCME.Types.ProhibitionVoltage
         public TypeManagement TypeManagement { get; set; }
 
         [DataMember]
-        public double ControlVoltage { get; set; }
+        public ushort ControlVoltage { get; set; }
         [DataMember]
-        public double ControlCurrent { get; set; }
+        public ushort ControlCurrent { get; set; }
 
         [DataMember]
-        public double SwitchedAmperage { get; set; }
+        public ushort SwitchedAmperage { get; set; }
         [DataMember]
-        public double SwitchedVoltage { get; set; }
+        public ushort SwitchedVoltage { get; set; }
 
         [DataMember]
-        public double AuxiliaryVoltagePowerSupply1 { get; set; }
+        public ushort AuxiliaryVoltagePowerSupply1 { get; set; }
         [DataMember]
-        public double AuxiliaryCurrentPowerSupply1 { get; set; }
+        public ushort AuxiliaryCurrentPowerSupply1 { get; set; }
 
         [DataMember]
-        public double AuxiliaryVoltagePowerSupply2 { get; set; }
+        public ushort AuxiliaryVoltagePowerSupply2 { get; set; }
         [DataMember]
-        public double AuxiliaryCurrentPowerSupply2 { get; set; }
+        public ushort AuxiliaryCurrentPowerSupply2 { get; set; }
 
         [DataMember]
         public double ProhibitionVoltageMinimum { get; set; }
